requirements.txt
fastapi==0.120.3
uvicorn==0.22.0
python-dotenv==1.0.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core
google-api-python-client
google-auth
protobuf
tqdm
